Move HSearch configuration tests to their own package
