+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/elasticsearch/test/configuration/NoConfigIndexedEntityTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.search.elasticsearch.test;
+package io.quarkus.hibernate.search.elasticsearch.test.configuration;
 
 import java.sql.SQLException;
 
@@ -7,16 +7,18 @@ import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
+import io.quarkus.deployment.configuration.ConfigurationError;
 import io.quarkus.test.QuarkusUnitTest;
 
-public class NoConfigNoIndexedEntityTest {
+public class NoConfigIndexedEntityTest {
 
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest().setArchiveProducer(
-            () -> ShrinkWrap.create(JavaArchive.class));
+            () -> ShrinkWrap.create(JavaArchive.class).addClass(IndexedEntity.class))
+            .setExpectedException(ConfigurationError.class);
 
     @Test
     public void testNoConfig() throws SQLException {
-        // we should be able to start the application, even with no configuration at all nor indexed entities
+        // an exception should be thrown
     }
 }
