+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/elasticsearch/test/configuration/NoConfigNoIndexedEntityTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.search.elasticsearch.test;
+package io.quarkus.hibernate.search.elasticsearch.test.configuration;
 
 import java.sql.SQLException;
 
@@ -7,18 +7,16 @@ import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import io.quarkus.deployment.configuration.ConfigurationError;
 import io.quarkus.test.QuarkusUnitTest;
 
-public class NoConfigIndexedEntityTest {
+public class NoConfigNoIndexedEntityTest {
 
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest().setArchiveProducer(
-            () -> ShrinkWrap.create(JavaArchive.class).addClass(IndexedEntity.class))
-            .setExpectedException(ConfigurationError.class);
+            () -> ShrinkWrap.create(JavaArchive.class));
 
     @Test
     public void testNoConfig() throws SQLException {
-        // an exception should be thrown
+        // we should be able to start the application, even with no configuration at all nor indexed entities
     }
 }
